pref: change id type from long to uint

+++ src/SweetLog.Serilog/Batching/LogBatch.cs
@@ -2,7 +2,7 @@ namespace SweetLog.Serilog.Batching;
 
 public struct LogBatch
 {
-    public long Id;
+    public uint Id;
 
     public long Timestamp;
     public int LogsCount;
